+++ example/linux_system_info_example.dart
@@ -1,6 +1,6 @@
 import 'package:linux_system_info/linux_system_info.dart';
 
-void main() {
+Future<void> main() async {
   //CPU
 
   var cpu_usage = CpuInfo
@@ -20,6 +20,11 @@ void main() {
   var total_swap = MemInfo()
       .swap_total_gb; // This returns the amount of SWAP in GB (you can also get it in kb or mb) e.g. 2
 
+  //GPU
+
+  var gpu_model = (await GpuInfo.load())[0]
+      .model; // This returns the model name of the GPU e.g. "Intel HD Graphics"
+
   //SYSTEM
 
   var kernel_name =
@@ -36,4 +41,10 @@ void main() {
   var os_release = SystemInfo()
       .os_release; // This value is a Map containing a parsed version of /etc/os-release
 
+  //GNOME
+  var gnome_version =
+      GnomeInfo().version; // This returns the Gnome version; e.g. '3.38.5'
+
+  var gnome_distributor = GnomeInfo()
+      .distributor; // This returns the Gnome distributor; e.g. 'Ubuntu'
 }