+++ lib/src/linux_system_info_base.dart
@@ -1,5 +1,11 @@
+import 'dart:ffi';
 import 'dart:io';
+import 'package:dbus/dbus.dart';
+import 'package:ffi/ffi.dart';
+import 'package:path/path.dart' as p;
 import 'package:tuple/tuple.dart';
+import 'package:xdg_directories/xdg_directories.dart' as xdg;
+import 'package:xml/xml.dart';
 
 /// Class for handling general system information
 class SystemInfo {
@@ -30,8 +36,8 @@ class SystemInfo {
       os_release[data[0]] = data[1].replaceAll('\"', '');
     }
 
-    os_name = os_release['NAME'];
-    os_version = os_release['VERSION_ID'];
+    os_name = os_release['NAME'] ?? '';
+    os_version = os_release['VERSION_ID'] ?? '';
   }
 }
 
@@ -436,5 +442,143 @@ class CpuInfo {
 
       sleep(Duration(seconds: 1));
     }
+  }
+}
+
+class Gpu {
+  /// The model name of the GPU e.g. "Intel(R) HD Graphics"
+  String model = '';
+
+  Gpu(this.model);
+}
+
+/// GPU information
+class GpuInfo {
+  /// Fetches GPU information from:
+  /// - `net.hadess.SwitcherooControl.GPUs` (D-Bus)
+  /// - `org.gnome.SessionManager.Renderer` (D-Bus)
+  /// - `GL_RENDERER` (OpenGL)
+  static Future<List<Gpu>> load() async {
+    final gpus = await _getSwitcherooGpus() ??
+        await _getGnomeRenderer() ??
+        await _getGlRenderer();
+    return gpus?.map((gpu) => Gpu(gpu)).toList() ?? [];
+  }
+
+  static Future<DBusValue?> _getDBusProperty({
+    required DBusClient client,
+    required String interface,
+    required String path,
+    required String property,
+  }) async {
+    try {
+      final object = DBusRemoteObject(
+        client,
+        name: interface,
+        path: DBusObjectPath(path),
+      );
+      final value = await object.getProperty(interface, property);
+      await client.close();
+      return value;
+    } on DBusErrorException catch (_) {
+      return null;
+    }
+  }
+
+  static Future<Iterable<String>?> _getSwitcherooGpus() async {
+    final value = await _getDBusProperty(
+      client: DBusClient.system(),
+      interface: 'net.hadess.SwitcherooControl',
+      path: '/net/hadess/SwitcherooControl',
+      property: 'GPUs',
+    );
+    if (value == null) return null;
+
+    final gpus =
+        (value as DBusArray).children.map((child) => child.toNative()).toList();
+    gpus.sort((a, b) {
+      if (a['Default'] == true) return 1;
+      if (b['Default'] == true) return -1;
+      return 0;
+    });
+    return gpus.map((gpu) => gpu['Name'] as String);
+  }
+
+  static Future<Iterable<String>?> _getGnomeRenderer() async {
+    final value = await _getDBusProperty(
+      client: DBusClient.system(),
+      interface: 'org.gnome.SessionManager',
+      path: '/org/gnome/SessionManager',
+      property: 'Renderer',
+    );
+    if (value == null) return null;
+    return <String>[(value as DBusString).value];
+  }
+
+  static Future<Iterable<String>?> _getGlRenderer() async {
+    try {
+      final lib = DynamicLibrary.open('libGL.so.1');
+      final glGetString =
+          lib.lookupFunction<_GlGetStringC, _GlGetStringDart>('glGetString');
+      const GL_RENDERER = 0x1F01;
+      final cstr = glGetString(GL_RENDERER);
+      if (cstr == nullptr) return null;
+      return <String>[cstr.cast<Utf8>().toDartString()];
+    } on ArgumentError catch (_) {
+      return null;
+    }
+  }
+}
+
+typedef _GlGetStringC = Pointer<Uint8> Function(Uint32 name);
+typedef _GlGetStringDart = Pointer<Uint8> Function(int name);
+
+class GnomeInfo {
+  GnomeInfo._([this.version = '', this.distributor = '']);
+
+  final String version;
+  final String distributor;
+
+  static GnomeInfo? _instance;
+
+  factory GnomeInfo() {
+    _instance ??= _init();
+    return _instance!;
+  }
+
+  static GnomeInfo _init() {
+    final file = _findDataFile('gnome/gnome-version.xml');
+    if (file == null) return GnomeInfo._();
+    final xml = file.readAsStringSync();
+    return _parseGnomeVersion(xml);
+  }
+
+  static GnomeInfo _parseGnomeVersion(String xml) {
+    final doc = XmlDocument.parse(xml);
+
+    final elements = <String, String>{
+      for (final element in doc.rootElement.childElements)
+        element.name.toString(): element.text,
+    };
+
+    final version = <String?>[
+      elements['platform'],
+      elements['minor'],
+      elements['micro'],
+    ].takeWhile((v) => v != null).join('.');
+
+    final distro = elements['distributor'];
+
+    return GnomeInfo._(version, distro ?? '');
+  }
+
+  static File? _findDataFile(String path) {
+    for (final dir in xdg.dataDirs) {
+      final file = File(p.join(dir.path, path));
+      if (file.existsSync()) {
+        return file;
+      }
+    }
+    return null;
   }
 }