+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.7
+
+- Added GnomeInfo and GPUInfo
+
+## 0.0.6
+
+- Minor fixes
+
+## 0.0.5
+
+- Moved to null safety
+
 ## 0.0.4
 
 - Formatted code with `dart format`